#126 [REFACTOR] import alias 설정 및 hooks 디렉터리 구조 개선
