+++ src/pages/MyPage/MyPage/PolicyTab.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Icon from '../../../components/Icon/Icon';
 import styles from './MyPage.module.css';
 import { Link } from 'react-router-dom';
-import useAuth from '../../../hooks/useAuth';
+import { useAuth } from '@/hooks';
 
 const PolicyTab = () => {
   const { logout } = useAuth();
